Revert "💄 design: ProjectTextEdit/#34"
