+++ src/components/Icons.tsx
@@ -80,19 +80,3 @@ export const EditIcon = ({
     />
   </svg>
 )
-
-export const CloseIcon = ({ fillColor = '#707070' }: iconProps) => (
-  <svg
-    width='24'
-    height='25'
-    viewBox='0 0 24 25'
-    fill='none'
-    xmlns='http://www.w3.org/2000/svg'
-  >
-    <rect y='0.5' width='24' height='24' rx='12' fill={fillColor} />
-    <path
-      d='M16.5303 9.03033C16.8232 8.73744 16.8232 8.26256 16.5303 7.96967C16.2374 7.67678 15.7626 7.67678 15.4697 7.96967L16.5303 9.03033ZM7.46967 15.9697C7.17678 16.2626 7.17678 16.7374 7.46967 17.0303C7.76256 17.3232 8.23744 17.3232 8.53033 17.0303L7.46967 15.9697ZM8.53034 7.96967C8.23745 7.67678 7.76257 7.67678 7.46968 7.96967C7.17678 8.26256 7.17678 8.73744 7.46968 9.03033L8.53034 7.96967ZM15.4697 17.0303C15.7626 17.3232 16.2374 17.3232 16.5303 17.0303C16.8232 16.7374 16.8232 16.2626 16.5303 15.9697L15.4697 17.0303ZM15.4697 7.96967L7.46967 15.9697L8.53033 17.0303L16.5303 9.03033L15.4697 7.96967ZM7.46968 9.03033L15.4697 17.0303L16.5303 15.9697L8.53034 7.96967L7.46968 9.03033Z'
-      fill='white'
-    />
-  </svg>
-)
