+++ src/components/index.ts
@@ -2,10 +2,6 @@
 export { TheHeader } from './layout/TheHeader'
 export { TheFooter } from './layout/TheFooter'
 export { ProjectHeader } from './layout/ProjectHeader'
-export { VoiceControlMenu } from './layout/VoiceControlMenu'
-export { TextEditList } from './layout/TextEditList'
-export { AvatarVoiceMenu } from './layout/AvatarVoiceMenu'
-export { ProjectText } from './layout/ProjectText'
 
 // ui
 
@@ -18,4 +14,3 @@ export { AvatarSelector } from './ui/AvatarSelector'
 export { ProjectList } from './ui/ProjectList'
 export { NextButton } from './ui/NextButton'
 export { VoiceUploadButton } from './ui/VoiceUploadButton'
-export { TextEnterButton } from './ui/TextEnterButton'
