--- src/components/layout/AvatarVoiceMenu/index.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export const AvatarVoiceMenu = () => {
-  return <div>AvatarVoiceMenu</div>
-}
