--- src/components/layout/ProjectText/index.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export const ProjectText = () => {
-  return <div>ProjectText</div>
-}
