--- src/components/layout/TextEditList/style.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components'
-
-export const Wrapper = styled.div`
-  width: 91.2rem;
-`
