--- src/components/ui/TextEnterButton/index.tsx
@@ -1,10 +1,0 @@
-import React, { useRef } from 'react'
-import * as S from './style'
-
-export const TextEnterButton = () => {
-  return (
-    <S.Button>
-      <span>+ 텍스트 입력</span>
-    </S.Button>
-  )
-}
