--- src/components/ui/TextEnterButton/style.ts
@@ -1,21 +1,0 @@
-import styled from 'styled-components'
-
-export const Button = styled.button`
-  width: 9.8rem;
-  height: 3.4rem;
-  background-color: #666666;
-  border-radius: 1rem;
-  border: none;
-  cursor: pointer;
-
-  span {
-    color: #ffffff;
-    font-size: 1.2rem;
-    font-weight: 500;
-    line-height: 2rem;
-  }
-
-  input {
-    display: none;
-  }
-`
