--- src/pages/ProjectTextEdit/style.ts
@@ -1,1 +1,0 @@
-import styled from 'styled-components'
