+++ src/pages/ProjectTextEnter/style.ts
@@ -1,15 +1,14 @@
 import styled from 'styled-components'
 
 export const Container = styled.div`
-  width: 38.4rem;
-  height: 71.2rem;
+  width: 97.6rem;
+  height: 20.1rem;
   background-color: #f0f0f0;
   border-radius: 1rem;
-  border: 1px solid blue;
 `
 export const Label = styled.div`
-  width: 100rem;
-  height: 100rem;
+  width: 21rem;
+  height: 2.5rem;
   color: #060606;
 
   font-weight: 500;
@@ -17,4 +16,6 @@ export const Label = styled.div`
   white-space: nowrap;
   /* 자간에 차이가 있는것인지 figma와 다르게 줄바꿈이 생김 */
 `
-export const ControllerContainer = styled.div``
+export const ControllerContainer = styled.div`
+  display: flex;
+`
